Move renderAskLocation to AskLocation module
